api/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
joblib==1.4.2
psycopg2-binary==2.9.10
numpy==2.2.4
xgboost==2.1.4
